CR @pepe: Use `next ds` for cycle (0 is not always the first position to get in a data structure)
This change also *removes* the automatic wrapping of a scalar data
term in an array (ie, being able to (cycle :ok) and have it "do the
right thing"). For the same reason that `mapcat` expects that the return
value of `f` be an iterable, we can't intelligently determine whether to
wrap `ds` or not in `cycle`. There's no way to inspect a data structure
and determine whether it's iterable or not, so we have to assume it is
and leave it to the user to pass in an iterable.