getline: Fix bugs with double-width chars
The cursor position actually needs to be tracked in three different
units: byte position within the buffer, for slicing and partial
printouts; screen column position, for determining overflow; and
character position, for cursor positioning leftward if not at end.

All three positions are cached and updated on-demand, so this should not
have a performance impact, though it does expose the real complexity of
things. And people think that TUIs are simpler than regular UIs : )