getline: Fix cursor positioning with double-width characters
There are two things going at confluence here.

a) The cursor positioning on VT terminals is based on columns. For every
   double-width character two "cursor forward" commands must be issued.
b) If the overflow is not even in respect to characters, i.e., the next
   character trimmed is double width because only one column is
   available, the cursor needs to be positioned not at the right margin
   of the screen, but one column leftward.

Thus, if the buffer is narrower than the screen width after trimming
overflow, we additionally move the cursor forward one fewer time to
match.

Some jumping of the cursor might still be present if single-width and
double-width characters are intermixed, given that the trimming will not
necessarily be linear on the left-hand-side; the narrowing algorithm is
conservative in this regard.