Change how dependencies are read from project.janet.
This fixes how we handle `janet-pm deps`, among other commands.
Previously, we would sometimes treat the `:depenencies` key as a list of
installed bundles, and other times as a list of package URIs. We should
instead commit to :dependencies in project.janet being a list of package
URIs, like it is for `jpm`. This is now a different meaning that in
info.jdn.

Instead of totally changing