Brief: add more pre/post steps for janet-pm
1. Adding these extra steps allows more user control of the build process
for complicated builds, including pre/post steps for build, install,
test and clean.

This also fixes #225, so that prebuild is only run before build.

2. Also removes redundant calls to build in the install hook since
Janet's core install hook calls build first as well. Having this in
spork too resulted in two calls to build when `janet-pm install`
was run.

3. Adds local assertf in `janet-pm` for older Janet

4. Fixes existing problem in `janet-pm` where hardcoded syspath at the top-level
is not honored when run as `(defn main ...`. This change adds a copy of the
hardcoded paths inside defn main in a binscript, if main exists in the input.