Brief: Fix missing dynamic :syspath inside binscript main.
Remove need for PEG for defn main in binscripts.

Summary: Rather than try to add code into a main in a binscript,
this change removes the extra line added to all binscripts
that remove the dynamic :syspath, if it was added at the top.

I'm not sure why it needs to be removed but certainly, removing
it means it is not defined when main runs which means janet-pm
tries to install into the system syspath, even when a dynamic
syspath is set.