Apply :has-bin-script less widely
:has-bin-script was being applied in too many
declare-* constructs so this has been dialed
back.

:has-exe has been added for the same purpose
with the idea that :has-bin-script will be
removed at some later point.  The additional
different name should help a bit with some of
the points brought up in #1648 in the janet
repository.

The plan is to have a companion change to
janet so both janet and spork will support
:has-bin-script and :has-exe for a while.