netrepl: allow specifying unix socket on CLI
Currently it is impossible to connect to or create a unix socket netrepl
server using the janet-netrepl CLI tool. It is however easy and useful
to create a unix socket based server using the (netrepl/server) function
directly.

This patch adds a new --unix-socket/-U <PATH> option that overrides any
host/port specified and connects to/creates a server at the given path.