fix http: Support other stream types, fix HEAD
It's useful to be able to use http with anything that exposes a
stream-like API, so we've changed a few things to make this
possible (mainly stopped using `ev/*` instead of methods like `:write`)

Forgot the `ev/chunk` calls in the previous commit.

Also fixed HEAD support (was inappropriately trying to read body when
the method was head, both with and without the changes).